requirements.txt
blinker==1.9.0
click==8.1.8
Flask==2.2.3
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pillow==11.2.1
Werkzeug==2.2.3
zipp==3.21.0
gunicorn==19.9.0
torch==2.1.0
torchvision==0.16.0
numpy<2
